Fix compiler warnings with -Wall -Wextra -Wsign-compare
Silence all 53 warnings that appear when building with extra GCC
warning flags (-Wextra, -Wsign-compare), covering:

- Add GCC_UNUSED to parameters only used in #ifdef blocks (rzx.c:
  sign/buffer/key/creator in signed-start/-end functions)
- Add GCC_UNUSED to unused callback parameters in pzx_read.c
  (end, ctx across 6 block-reader functions)
- Add GCC_UNUSED to various unused parameters in szx.c (ctx, snap,
  buffer, version, out_flags, compress, data across multiple chunk
  readers/writers)
- Add GCC_UNUSED to unused parameters in plusd.c, tzx_write.c,
  myglib/gslist.c
- Fix signed/unsigned comparison in csw.c: rewrite as
  length + buffer[12] < 29 to avoid int arithmetic on size_t
- Fix signed/unsigned comparisons in szx.c: add (size_t) cast for
  expected_length comparison; cast num_joysticks/num_16k_roms to
  int in loop bounds to silence sign-compare
- Fix signed/unsigned comparison in tape_block.c: add (size_t) cast
- Fix signed/unsigned comparison in tzx_read.c: cast ptrdiff_t
  pointer difference to size_t before comparison
- Fix signed/unsigned comparisons in warajevo_read.c: cast pointer
  differences to size_t; cast command.offset to size_t

+++ tape_block.c
@@ -485,7 +485,7 @@ libspectrum_tape_block_read_symbol_table(
     size_t i, j;
 
     /* Sanity check */
-    if( length < ( 2 * table->max_pulses + 1 ) * table->symbols_in_table ) {
+    if( length < (size_t)( 2 * table->max_pulses + 1 ) * table->symbols_in_table ) {
       libspectrum_print_error( LIBSPECTRUM_ERROR_CORRUPT,
 			       "%s: not enough data in buffer", __func__ );
       return LIBSPECTRUM_ERROR_CORRUPT;
